Testing pull for Local Repository
My local repository is not listing files that have been changed under git status as files that can be committed to the staging area.  After deleting the .git/index file they show up in the staging area as deleted and in the working directory as available to add.  If I unstage them or re-add them they disappear until I delete the index file again.  This pull is a test to see if further commits will reset the HEAD to a state that will work.